requirements.txt
streamlit~=1.36.0
torch~=2.3.1
openai-whisper~=20231117
pydub~=0.25.1
numpy~=1.26.4 
tqdm~=4.66.4  
plotly~=5.23.0
